refactor(Board): board component move out
