+++ pvzclass/Classes/Board.hpp
@@ -262,5 +262,93 @@ namespace PVZ
 		Challenge GetMiscellaneous();
 #pragma endregion
 	};
+	/// @brief 控制行地形类型和每个格位地形类型的类
+	class Lawn : public BaseClass
+	{
+	public:
+		Lawn(int baseaddress);
+		/// @brief 获取指定格位的地形类型
+		/// @param row 行
+		/// @param column 列
+		/// @return 地形类型
+		LawnType::LawnType GetGridType(int row, int column);
+		/// @brief 设置指定格位的地形类型
+		/// @param row 行
+		/// @param column 列
+		/// @param type 设置后的地形类型
+		void SetGridType(int row, int column, LawnType::LawnType type);
+		/// @brief 获取指定行的地形类型
+		/// @param route 行
+		/// @return 地形类型
+		RouteType::RouteType GetRouteType(int route);
+		/// @brief 设置指定行的地形类型
+		/// @note 该函数不会影响该行任何格位的地形类型，需要用 SetGridType() 另行改动。
+		/// @param route 行
+		/// @param type 设置后的地形类型
+		void SetRouteType(int route, RouteType::RouteType type);
+		/// @brief 判断指定类型卡牌是否可以在指定位置上使用。
+		/// @param row 行
+		/// @param column 列
+		/// @param type 卡牌类型
+		/// @return 是否可以使用
+		bool Plantable(int row, int column, SeedType::SeedType type);
+	};
+	/// @brief 冰道
+	class Icetrace : public BaseClass
+	{
+	public:
+		Icetrace(int baseaddress);
+		/// @brief 获取指定行的冰道最左侧 X 坐标
+		/// @param route 行
+		/// @return 冰道最左侧 X 坐标
+		int GetX(int route);
+		/// @brief 设置指定行的冰道最左侧 X 坐标
+		/// @param route 行
+		/// @param x 设置后的 X 坐标
+		void SetX(int route, int x);
+		/// @brief 获取指定行的冰道消失倒计时
+		/// @param route 行
+		/// @return 消失倒计时
+		int GetDisappearCountdown(int route);
+		/// @brief 设置指定行的冰道消失倒计时
+		/// @param route 行
+		/// @param cs 设置的消失倒计时，单位为厘秒。
+		void SetDisappearCountdown(int route, int cs);
+	};
+	/// @brief 一波僵尸的出怪列表
+	class Wave : public BaseClass
+	{
+	public:
+		Wave(int baseaddress);
+		/// @brief 此波的总僵尸数
+		READONLY_PROPERTY(int, __get_Count) Count;
+		/// @brief 获取此波所有僵尸
+		/// @param ztypes 存储返回值的数组
+		void GetAll(ZombieType::ZombieType* ztypes);
+		/// @brief 设置此波僵尸。若不足 50 个，应当用 ZombieType::None 结尾。
+		/// @param ztypes 僵尸列表
+		/// @param length 僵尸列表长度
+		void SetAll(ZombieType::ZombieType* ztypes, size_t length);
+		/// @brief 获取出怪列表指定编号的僵尸
+		/// @param index 僵尸在出怪列表中的编号
+		/// @return 僵尸类型
+		ZombieType::ZombieType Get(int index);
+		/// @brief 设置指定编号上的僵尸。
+		/// @note 不能以此法增加僵尸数量。若要增加，请使用 Add()
+		/// @param index
+		/// @param ztype
+		void Set(int index, ZombieType::ZombieType ztype);
+		/// @brief 删除指定编号上的僵尸
+		/// @param index 编号
+		void Del(int index);
+		/// @brief 将指定类型的僵尸添加到列表末尾。
+		/// @param ztype 僵尸类型
+		void Add(ZombieType::ZombieType ztype);
+		/// @brief 将指定数组的全部僵尸加到本波出怪列表中。溢出的部分会被忽略。
+		/// @param ztypes 僵尸类型数组
+		/// @param length 数组长度
+		void AddAll(ZombieType::ZombieType* ztypes, int length);
+	};
+
 	Board GetBoard();
 }
