+++ pvzclass/Classes/Icetrace.cpp
@@ -1,4 +1,4 @@
-﻿#include "../PVZ.h"
+#include "Board.hpp"
 
 PVZ::Icetrace::Icetrace(int baseaddress)
 {
