+++ pvzclass/Classes/Wave.cpp
@@ -1,4 +1,4 @@
-﻿#include "../PVZ.h"
+#include "Board.hpp"
 
 PVZ::Wave::Wave(int baseaddress)
 {
